id-card-service/pyproject.toml
[project]
name = "id-card-service"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dannhanong", email = "thannong0512@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "paddleocr>=3.0.1",
    "Pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "setuptools>=80.9.0",
    "paddlepaddle>=3.0.0",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "psutil>=7.0.0",
    "openai>=1.63.2",
    "py-eureka-client>=0.11.13",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
